Code/GraphMol/ChemReactions/ReactionFingerprints.cpp Code/GraphMol/Descriptors/AUTOCORR2D.cpp Code/GraphMol/Descriptors/ConnectivityDescriptors.cpp Code/GraphMol/Descriptors/DCLV.cpp Code/GraphMol/Descriptors/EEM.cpp Code/GraphMol/Descriptors/GETAWAY.cpp Code/GraphMol/Descriptors/WHIM.cpp Code/GraphMol/FilterCatalog/FilterCatalogRunner.cpp Code/GraphMol/Fingerprints/FingerprintGenerator.cpp Code/GraphMol/Fingerprints/MHFP.h

+++ Code/RDGeneral/testDict.cpp
@@ -96,7 +96,7 @@ void testRDAny() {
     bool a = true;
     RDValue v(a);
     TEST_ASSERT(rdvalue_cast<bool>(v) == true);
-    v = (int)10;
+    v = 10;
     TEST_ASSERT(rdvalue_cast<int>(v) == 10);
   }
 
@@ -195,7 +195,7 @@ void testRDAny() {
     std::clock_t clock2 = std::clock();
 
     std::cout << "static boost any:"
-              << (double)(clock2 - clock1) / CLOCKS_PER_SEC << " s"
+              << static_cast<double>(clock2 - clock1) / CLOCKS_PER_SEC << " s"
               << std::endl;
   }
   {
@@ -210,7 +210,7 @@ void testRDAny() {
     std::clock_t clock2 = std::clock();
 
     std::cout << "dynamic boost any:"
-              << (double)(clock2 - clock1) / CLOCKS_PER_SEC << " s"
+              << static_cast<double>(clock2 - clock1) / CLOCKS_PER_SEC << " s"
               << std::endl;
   }
 
@@ -222,8 +222,9 @@ void testRDAny() {
     }
     std::clock_t clock2 = std::clock();
 
-    std::cout << "static RDAny:" << (double)(clock2 - clock1) / CLOCKS_PER_SEC
-              << " s" << std::endl;
+    std::cout << "static RDAny:"
+              << static_cast<double>(clock2 - clock1) / CLOCKS_PER_SEC << " s"
+              << std::endl;
   }
 
   {
@@ -237,8 +238,9 @@ void testRDAny() {
     delete vv;
     std::clock_t clock2 = std::clock();
 
-    std::cout << "dynamic RDAny:" << (double)(clock2 - clock1) / CLOCKS_PER_SEC
-              << " s" << std::endl;
+    std::cout << "dynamic RDAny:"
+              << static_cast<double>(clock2 - clock1) / CLOCKS_PER_SEC << " s"
+              << std::endl;
   }
 
   {
@@ -249,8 +251,9 @@ void testRDAny() {
     }
     std::clock_t clock2 = std::clock();
 
-    std::cout << "static RDValue:" << (double)(clock2 - clock1) / CLOCKS_PER_SEC
-              << " s" << std::endl;
+    std::cout << "static RDValue:"
+              << static_cast<double>(clock2 - clock1) / CLOCKS_PER_SEC << " s"
+              << std::endl;
   }
 
   {
@@ -263,7 +266,7 @@ void testRDAny() {
     std::clock_t clock2 = std::clock();
 
     std::cout << "dynamic RDValue:"
-              << (double)(clock2 - clock1) / CLOCKS_PER_SEC << " s"
+              << static_cast<double>(clock2 - clock1) / CLOCKS_PER_SEC << " s"
               << std::endl;
   }
 
@@ -530,8 +533,8 @@ void testConstReturns() {
     }
     end = std::clock();
     BOOST_LOG(rdErrorLog) << "done: "
-                          << (end - start) / (double)(CLOCKS_PER_SEC) << " "
-                          << ls << std::endl;
+                          << (end - start) / static_cast<double>(CLOCKS_PER_SEC)
+                          << " " << ls << std::endl;
 
     ls = 0;
     BOOST_LOG(rdErrorLog) << "copy" << std::endl;
@@ -542,8 +545,8 @@ void testConstReturns() {
     }
     end = std::clock();
     BOOST_LOG(rdErrorLog) << "done: "
-                          << (end - start) / (double)(CLOCKS_PER_SEC) << " "
-                          << ls << std::endl;
+                          << (end - start) / static_cast<double>(CLOCKS_PER_SEC)
+                          << " " << ls << std::endl;
 
     ls = 0;
     BOOST_LOG(rdErrorLog) << "ref" << std::endl;
@@ -554,8 +557,8 @@ void testConstReturns() {
     }
     end = std::clock();
     BOOST_LOG(rdErrorLog) << "done: "
-                          << (end - start) / (double)(CLOCKS_PER_SEC) << " "
-                          << ls << std::endl;
+                          << (end - start) / static_cast<double>(CLOCKS_PER_SEC)
+                          << " " << ls << std::endl;
 
     ls = 0;
     BOOST_LOG(rdErrorLog) << "dict" << std::endl;
@@ -566,8 +569,8 @@ void testConstReturns() {
     }
     end = std::clock();
     BOOST_LOG(rdErrorLog) << "done: "
-                          << (end - start) / (double)(CLOCKS_PER_SEC) << " "
-                          << ls << std::endl;
+                          << (end - start) / static_cast<double>(CLOCKS_PER_SEC)
+                          << " " << ls << std::endl;
 
     ls = 0;
     BOOST_LOG(rdErrorLog) << "ref with hasVal" << std::endl;
@@ -581,8 +584,8 @@ void testConstReturns() {
     }
     end = std::clock();
     BOOST_LOG(rdErrorLog) << "done: "
-                          << (end - start) / (double)(CLOCKS_PER_SEC) << " "
-                          << ls << std::endl;
+                          << (end - start) / static_cast<double>(CLOCKS_PER_SEC)
+                          << " " << ls << std::endl;
 
     // std::string nv=d.getVal<std::string>("foo");
   }
